@@ -25,9 +25,6 @@ return [
     // file storage service, such as s3, to store publicly accessible assets.
     'url' => env('STORAGE_URL', false),
 
-    // Default Cloud Filesystem Disk
-    'cloud' => 's3',
-
     // Available filesystem disks
     // Only local, local_secure & s3 are supported by BookStack
     'disks' => [
@@ -35,6 +32,7 @@ return [
         'local' => [
             'driver' => 'local',
             'root'   => public_path(),
+            'visibility' => 'public',
         ],
 
         'local_secure_attachments' => [
@@ -45,6 +43,7 @@ return [
         'local_secure_images' => [
             'driver' => 'local',
             'root'   => storage_path('uploads/images/'),
+            'visibility' => 'public',
         ],
 
         's3' => [
@@ -57,6 +56,14 @@ return [
             'use_path_style_endpoint' => env('STORAGE_S3_ENDPOINT', null) !== null,
         ],
 
+    ],
+
+    // Symbolic Links
+    // Here you may configure the symbolic links that will be created when the
+    // `storage:link` Artisan command is executed. The array keys should be
+    // the locations of the links and the values should be their targets.
+    'links' => [
+        public_path('storage') => storage_path('app/public'),
     ],
 
 ];
