@@ -12,7 +12,7 @@ class SortTest extends TestCase
 {
     protected $book;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->book = Book::first();
