@@ -23,7 +23,7 @@ class EntityPermissionsTest extends TestCase
      */
     protected $viewer;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
         $this->user = $this->getEditor();
